finished router class refactor
