+++ packages/stacked_generator/lib/src/generators/router/generator/routes_class/routes_class_builder.dart
