+++ packages/stacked_generator/lib/src/generators/router/route_config/adaptive_route_config.dart
@@ -21,7 +21,6 @@ class AdaptiveRouteConfig extends RouteConfig {
   @override
   String registerRoutes() {
     StringBuffer stringBuffer = StringBuffer();
-    stringBuffer.write(super.registerArgs());
     stringBuffer.write(
         'return buildAdaptivePageRoute<$processedReturnType>(builder: (context) => $joinedConstructerParams, settings: data,');
     if (cupertinoNavTitle != null) {
