+++ packages/stacked_generator/test/router/router_class/router_class_builder_test.dart
@@ -1,4 +1,3 @@
-import 'package:code_builder/code_builder.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
 import 'package:stacked_generator/src/generators/router/generator/router_class/router_class_builder.dart';
 import 'package:stacked_generator/src/generators/router/route_config/material_route_config.dart';
