+++ packages/stacked_generator/test/router/routes_class_builder_test.dart
@@ -1,6 +1,6 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
-import 'package:stacked_generator/src/generators/router/generator/routes_class_builder.dart';
+import 'package:stacked_generator/src/generators/router/generator/routes_class/routes_class_builder.dart';
 import 'package:stacked_generator/src/generators/router/route_config/material_route_config.dart';
 import 'package:test/test.dart';
